+++ src/gui/menu.hpp
@@ -46,7 +46,7 @@ public:
   MenuItem* add_submenu(const std::string& text, int submenu, int id = -1);
   MenuItem* add_controlfield(int id, const std::string& text,
                              const std::string& mapping = "");
-  MenuItem* add_string_select(int id, const std::string& text, size_t* selected, std::vector<std::string> strings);
+  MenuItem* add_string_select(int id, const std::string& text, int* selected, std::vector<std::string> strings);
   MenuItem* add_textfield(const std::string& text, std::string* input, int id = -1);
   MenuItem* add_intfield(const std::string& text, int* input, int id = -1);
   MenuItem* add_numfield(const std::string& text, float* input, int id = -1);
